refactor: move constants and embedded config templates into oarepo_cli.configuration
- Moves oarepo_cli/constants.py to oarepo_cli/configuration/constants.py
  and updates its three importers.
- Replaces the RUFF_TOML/MYPY_INI string constants in lint.py with real
  data files (ruff.toml.tmpl, mypy.ini.tmpl) bundled in the new package,
  loaded lazily via importlib.resources (oarepo_cli.configuration.resources
  .read_text()) only when a command actually needs to write them out -
  not at import time.
- Templates use a .tmpl suffix so ruff doesn't auto-discover a bare
  ruff.toml here as its own config when linting oarepo-cli's own source
  under this package (verified this was silently happening before the
  rename).
- Verified both template files land in the built wheel.
