+++ oarepo_cli/configuration/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: 2026 CESNET z.s.p.o.
+# SPDX-License-Identifier: MIT
+
+"""Application-wide constants and bundled configuration template files."""
+
+from __future__ import annotations
