+++ oarepo_cli/configuration/mypy.ini.tmpl
@@ -1,0 +1,5 @@
+[mypy]
+warn_return_any = True
+warn_unused_configs = True
+warn_unreachable = True
+follow_untyped_imports = True
