+++ oarepo_cli/configuration/ruff.toml.tmpl
@@ -1,0 +1,45 @@
+target-version = "py314"
+line-length = 120
+indent-width = 4
+
+[lint]
+select = [ "ALL" ]
+ignore = [
+    "FIX002",  # exclude TODO comments
+    "TD003",   # Missing issue link for this TODO
+    "TD002",   # Missing author in TODO
+    "N806",    # Variable name should be lowercase as we dynamically create classes
+    "ANN204",  # Missing return type annotation for __init__
+    "ANN401",  # Any in *args/**kwargs
+    "TRY003",  # Avoid long exception messages
+    "EM101",   # Avoid using string literal in exception
+    "EM102",   # Avoid using f-string literal in exception
+    "TRY301",  # Avoid raising/catching the same exception type
+    "PLC0415",  # Place imports to the top of the file
+    "PGH004",  # Use specific noqa for pylint
+    "TID252",  # Prefer absolute imports
+    "D203",    # Using D211
+    "D213",    # Using D212 (multi-line-summary-first-line) instead
+    "COM812",
+    "FBT001",  # Avoid using boolean function parameters
+    "FBT002",  # Avoid using boolean function parameters
+]
+
+[lint.per-file-ignores]
+"__init__.py" = ["E402"]
+"**/{tests,docs,tools}/*" = [
+    "E402",
+    "S101",
+    "ANN001",
+    "ARG001",
+    "D103",
+    "ANN201",
+    "D100",
+    "INP",
+    "PLR",
+    "PLC"
+    ]
+
+[format]
+docstring-code-format = true
+docstring-code-line-length = 40
